convert position Vec3 to Vec3d
* This improves physics calculations
* Fixes some bugs in extreme coordinates
* Hopefully fixes the no move bug on hypixel

+++ src/main/java/de/bixilon/minosoft/config/config/game/CameraGameConfig.kt
@@ -17,7 +17,7 @@ import com.squareup.moshi.Json
 
 data class CameraGameConfig(
     @Json(name = "render_distance") var renderDistance: Int = 10,
-    var fov: Float = 60f,
+    var fov: Double = 60.0,
     @Json(name = "dynamic_fov") var dynamicFov: Boolean = true,
     @Json(name = "mouse_sensitivity") var moseSensitivity: Float = 0.1f,
     @Json(name = "no_clip_movement") var noCipMovement: Boolean = false,
